+++ What to do
@@ -1,4 +1,4 @@
-#Make sure to inform in whatsapp group which part you are doing.
+# Make sure to inform in whatsapp group which part you are doing.
 
 [DONE] 1) classes for Customer , Cars , rental table. Cars class being abstract.
 
@@ -8,13 +8,13 @@
 [DONE] 2)(b) Function to make a booking using Cust_id, Car_no.,start (date), end (date) 
     input format in cmd line will be   ->  java Main -booking <Cust_id> <Car_no> <start> <end>
 
-2)(c) Function to cancel booking using Cust_id and Car_no. 
+[Done] 2)(c) Function to cancel booking using Cust_id and Car_no. 
     input format in cmd line will be   ->  java Main -cancel <Cust_id> <Car_no>
     
 [Done] 3) Function to search and print Customer data using first name.
     input format in cmd line will be   ->  java Main -searchFirstName <FirstName>
     
-4) Function to find available cars with Charges less than a certain amount. 
+[Done] 4) Function to find available cars with Charges less than a certain amount. 
   For example, if the user want to find avaible cars for renting which price is less that 1000 rs. per day
     input format in cmd line will be   ->  java Main -searchLessThanAmount <Amount>
