--- carData.csv
@@ -1,5 +1,0 @@
-AP-123,SUV,1,2000
-TS-123,HATCH BACK,1,1000
-AP-345,SEDAN,1,1500
-AP-78,SUV,1,2000
-AP-76,SEDAN,1,1600
