+++ package1/customer.java
@@ -1,0 +1,14 @@
+package package1;
+public class customer{
+    int cust_id;
+    String name;
+    String address;
+    String phone_no;
+
+    public customer(int cust_id, String name, String address, String phone_no) {
+        this.cust_id = cust_id;
+        this.name = name;
+        this.address = address;
+        this.phone_no = phone_no;
+    }
+}
