+++ package2/rental.java
@@ -1,0 +1,15 @@
+package package2;
+public class rental{
+    int cust_id;
+    String car_no;
+    String start;
+    String end;
+    int total_fee;
+
+    public rental(int cust_id, String car_no, String start, String end) {
+        this.cust_id = cust_id;
+        this.car_no = car_no;
+        this.start = start;
+        this.end = end;
+    }
+}
